--- a/BULK_SAVE_PERFORMANCE_OPTIMIZATION.md
+++ b/BULK_SAVE_PERFORMANCE_OPTIMIZATION.md
@@ -220,3 +220,5 @@ The optimizations are implemented and ready for testing:
 
 
 
+
+