--- a/BULK_SAVE_PERFORMANCE_OPTIMIZATION.md
+++ b/BULK_SAVE_PERFORMANCE_OPTIMIZATION.md
@@ -222,3 +222,4 @@ The optimizations are implemented and ready for testing:
 
 
 
+