--- a/SMART_DEDUPLICATION_IMPACT_AND_CHANGES.md
+++ b/SMART_DEDUPLICATION_IMPACT_AND_CHANGES.md
@@ -209,3 +209,5 @@ Let me know if you need help with the updates!
 
 
 
+
+