--- a/SMART_DEDUPLICATION_IMPACT_AND_CHANGES.md
+++ b/SMART_DEDUPLICATION_IMPACT_AND_CHANGES.md
@@ -211,3 +211,4 @@ Let me know if you need help with the updates!
 
 
 
+