--- a/composer.json
+++ b/composer.json
@@ -77,7 +77,6 @@
 		"guzzlehttp/guzzle": "^7.0",
 		"opis/json-schema": "^2.3",
 		"phpoffice/phpspreadsheet": "^4.2",
-		"react/async": "^4.3",
 		"react/event-loop": "^1.5",
 		"react/promise": "^3.2",
 		"symfony/uid": "^6.4",