--- a/lib/Controller/HeartbeatController.php
+++ b/lib/Controller/HeartbeatController.php
@@ -70,3 +70,5 @@ class HeartbeatController extends Controller
 
 
 
+
+