--- a/lib/Controller/HeartbeatController.php
+++ b/lib/Controller/HeartbeatController.php
@@ -72,3 +72,4 @@ class HeartbeatController extends Controller
 
 
 
+