--- a/lib/Settings/softwarecatalogus_register.json
+++ b/lib/Settings/softwarecatalogus_register.json
@@ -2223,6 +2223,7 @@
                 "configuration": {
                     "objectNameField": "afnemer", 
                     "objectDescriptionField": "product",
+                    "objectSummaryField": "status",
                     "allowFiles": true,
                     "allowedTags": [
                         "DPIA",